lf review: docs: add maestro app, lfd daemon, and api reference
Added comprehensive documentation for Maestro macOS app, lfd daemon service, socket protocol API, and project philosophy. Added internal READMEs for lfd and Maestro modules.
