+++ src/loopflow/cli/run.py
@@ -117,7 +117,7 @@ def _execute_task(
     collector_cmd = [
         sys.executable,
         "-m",
-        "loopflow.maestro.collector",
+        "loopflow.lfd.collector",
         "--session-id",
         session.id,
         "--task",
